More accurately document normal form conversion functions.
The Haddock docs now correctly state which logical laws are
employed by the different conversion functions, and give an outline
in all cases of how the conversion process proceeds.

This is helpful since the code is not terribly transparent, as it
has to deal with many cases, so even if a particular line is clear
it's not always obvious what the algorithm is for arbitrary
expressions.